HPCC-30196 Move k8s helpers to jlib
