+++ esp/services/ws_smc/ws_smcService.cpp
@@ -17,6 +17,7 @@
 
 #pragma warning (disable : 4786)
 
+#include "jcontainerized.hpp"
 
 #ifdef _USE_OPENLDAP
 #include "ldapsecurity.ipp"
@@ -2258,7 +2259,7 @@ bool CWsSMCEx::onRoxieXrefCmd(IEspContext &context, IEspRoxieXrefCmdRequest &req
     ISmartSocketFactory *conn = roxieConnMap.getValue(target);
     if (!conn)
         throw makeStringExceptionV(ECLWATCH_CANNOT_GET_ENV_INFO, "roxie target cluster not mapped: %s", target);
-    if (!isActiveK8sService(target))
+    if (!k8s::isActiveService(target))
         throw makeStringExceptionV(ECLWATCH_CANNOT_GET_ENV_INFO, "roxie target cluster has no active servers: %s", target);
     Owned<IPropertyTree> controlResp;
     if (req.getCheckAllNodes())
@@ -2319,7 +2320,7 @@ bool CWsSMCEx::onRoxieControlCmd(IEspContext &context, IEspRoxieControlCmdReques
     ISmartSocketFactory *conn = roxieConnMap.getValue(target);
     if (!conn)
         throw makeStringExceptionV(ECLWATCH_CANNOT_GET_ENV_INFO, "roxie target cluster not mapped: %s", target);
-    if (!isActiveK8sService(target))
+    if (!k8s::isActiveService(target))
         throw makeStringExceptionV(ECLWATCH_CANNOT_GET_ENV_INFO, "roxie target cluster has no active servers: %s", target);
 
     Owned<IPropertyTree> controlResp = sendRoxieControlAllNodes(conn, controlReq, true, req.getWait(), ROXIECONNECTIONTIMEOUT);
