+++ esp/services/ws_workunits/ws_workunitsQuerySets.cpp
@@ -18,6 +18,7 @@
 #include "ws_workunitsService.hpp"
 #include "ws_fs.hpp"
 #include "jlib.hpp"
+#include "jcontainerized.hpp"
 #include "jflz.hpp"
 #include "daclient.hpp"
 #include "dalienv.hpp"
@@ -527,7 +528,7 @@ bool reloadCluster(MapStringToMyClass<ISmartSocketFactory> &roxieConnMap, const 
     ISmartSocketFactory *conn = roxieConnMap.getValue(target);
     if (!conn)
         return true;
-    if (!isActiveK8sService(target))
+    if (!k8s::isActiveService(target))
         return true;
 #endif
     try
@@ -849,7 +850,7 @@ bool CWsWorkunitsEx::isQuerySuspended(const char* query, const char* target, uns
         ISmartSocketFactory *conn = roxieConnMap.getValue(target);
         if (!conn)
             return false;
-        if (!isActiveK8sService(target))
+        if (!k8s::isActiveService(target))
             return false;
 #endif
         StringBuffer control;
@@ -1329,7 +1330,7 @@ IPropertyTree *getQueriesOnCluster(const char *target, const char *queryset, Str
     ISmartSocketFactory *conn = roxieConnMap.getValue(target);
     if (!conn)
         return nullptr;
-    if (!isActiveK8sService(target))
+    if (!k8s::isActiveService(target))
         return nullptr;
 #endif
 
@@ -1523,7 +1524,7 @@ unsigned CWsWorkunitsEx::getGraphIdsByQueryId(const char *target, const char *qu
     ISmartSocketFactory *conn = roxieConnMap.getValue(target);
     if (!conn)
         return 0;
-    if (!isActiveK8sService(target))
+    if (!k8s::isActiveService(target))
         return 0;
 #endif
 
@@ -3998,7 +3999,7 @@ bool CWsWorkunitsEx::onWUQueryGetSummaryStats(IEspContext& context, IEspWUQueryG
         ISmartSocketFactory *conn = roxieConnMap.getValue(target);
         if (!conn)
             throw makeStringException(ECLWATCH_INVALID_CLUSTER_NAME, "Roxie name not found");
-        if (!isActiveK8sService(target))
+        if (!k8s::isActiveService(target))
             throw makeStringException(ECLWATCH_INVALID_CLUSTER_NAME, "Roxie cluster has no active servers");
 #endif
