+++ thorlcr/master/thmastermain.cpp
@@ -30,6 +30,7 @@
 #endif
 
 #include "jlib.hpp"
+#include "jcontainerized.hpp"
 #include "jdebug.hpp"
 #include "jfile.hpp"
 #include "jmisc.hpp"
@@ -1020,11 +1021,11 @@ int main( int argc, const char *argv[]  )
             StringBuffer myEp;
             queryMyNode()->endpoint().getUrlStr(myEp);
 
-            workerNSInstalled = applyK8sYaml("thorworker", workunit, cloudJobName, "networkpolicy", { }, false, true);
+            workerNSInstalled = k8s::applyYaml("thorworker", workunit, cloudJobName, "networkpolicy", { }, false, true);
             if (workerNSInstalled)
             {
-                KeepK8sJobs keepJob = translateKeepJobs(globals->queryProp("@keepJobs"));
-                workerJobInstalled = applyK8sYaml("thorworker", workunit, cloudJobName, "job", { { "graphName", graphName}, { "master", myEp.str() }, { "_HPCC_NUM_WORKERS_", std::to_string(numWorkers/numWorkersPerPod)} }, false, KeepK8sJobs::none == keepJob);
+                k8s::KeepJobs keepJob = k8s::translateKeepJobs(globals->queryProp("@keepJobs"));
+                workerJobInstalled = k8s::applyYaml("thorworker", workunit, cloudJobName, "job", { { "graphName", graphName}, { "master", myEp.str() }, { "_HPCC_NUM_WORKERS_", std::to_string(numWorkers/numWorkersPerPod)} }, false, k8s::KeepJobs::none == keepJob);
                 if (workerJobInstalled)
                     doWorkerRegistration = true;
             }
@@ -1145,18 +1146,18 @@ int main( int argc, const char *argv[]  )
             {
                 try
                 {
-                    KeepK8sJobs keepJob = translateKeepJobs(globals->queryProp("@keepJobs"));
+                    k8s::KeepJobs keepJob = k8s::translateKeepJobs(globals->queryProp("@keepJobs"));
                     switch (keepJob)
                     {
-                        case KeepK8sJobs::all:
+                        case k8s::KeepJobs::all:
                             // do nothing
                             break;
-                        case KeepK8sJobs::podfailures:
+                        case k8s::KeepJobs::podfailures:
                             if (nullptr == exception)
-                                deleteK8sResource("thorworker", "job", cloudJobName);
+                                k8s::deleteResource("thorworker", "job", cloudJobName);
                             break;
-                        case KeepK8sJobs::none:
-                            deleteK8sResource("thorworker", "job", cloudJobName);
+                        case k8s::KeepJobs::none:
+                            k8s::deleteResource("thorworker", "job", cloudJobName);
                             break;
                     }
                 }
@@ -1170,7 +1171,7 @@ int main( int argc, const char *argv[]  )
             {
                 try
                 {
-                    deleteK8sResource("thorworker", "networkpolicy", cloudJobName);
+                    k8s::deleteResource("thorworker", "networkpolicy", cloudJobName);
                 }
                 catch (IException *e)
                 {
