+++ thorlcr/slave/thslavemain.cpp
@@ -26,6 +26,7 @@
 #include <stdlib.h> 
 
 #include "jlib.hpp"
+#include "jcontainerized.hpp"
 #include "jdebug.hpp"
 #include "jexcept.hpp"
 #include "jfile.hpp"
@@ -106,7 +107,7 @@ static bool RegisterSelf(SocketEndpoint &masterEp)
         CMessageBuffer msg;
         msg.append(mySlaveNum);
         if (isContainerized())
-            msg.append(queryMyPodName());
+            msg.append(k8s::queryMyPodName());
         queryWorldCommunicator().send(msg, masterNode, MPTAG_THORREGISTRATION);
         if (!queryWorldCommunicator().recv(msg, masterNode, MPTAG_THORREGISTRATION))
             return false;
